rename: `src/lib/api.ts` -> `src/api/post.ts`
