+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Headline from '@/components/Headline'
 import PostList from '@/components/PostList'
 import { siteDescription, siteTitle } from '@/config'
-import { getPosts } from '@/lib/api'
+import { getPosts } from '@/lib/post'
 
 const IndexPage: React.FC = async () => {
     const posts = await getPosts()
