+++ src/app/tags/[tag]/page.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Headline from '@/components/Headline'
 import PostList from '@/components/PostList'
 import { Tag as TagIcon } from '@/components/MaterialIcon'
-import { getPosts } from '@/lib/api'
+import { getPosts } from '@/lib/post'
 
 export type Params = {
     tag: string
